refactor: split to utils
