+++ src/utils/getSliderBreakpointsOptions.ts
@@ -1,0 +1,16 @@
+
+
+export const getSliderBreakpointsOptions = () => {
+    return{
+    320: {
+      slidesPerView: 1,
+      spaceBetween: 0,
+      initialSlide: 0,
+    },
+    1440: {
+      slidesPerView: 3,
+      spaceBetween: 20,
+      initialSlide: 1,
+    },
+}
+  }
